add optional uuid and dataset_uuids fields to Report model, with some refactoring to separate out UUID serialization
